Clean up #2
