+++ data/src/neo_processor.py
@@ -12,7 +12,7 @@ from .api_client import NASAAPIClient
 from .data_processor import NEODataProcessor as DataProcessor
 from .storage import DataLakeStorage
 from .models import ProcessingResult
-from .exceptions import NEOProcessorError, NASAAPIError, DataProcessingError, StorageError
+from .models import NEOProcessorError, NASAAPIError, DataProcessingError, StorageError
 
 logger = logging.getLogger(__name__)
